+++ packages/react-components/react-menu-grid-preview/library/src/components/MenuGridCell/useMenuGridCellContextValues.ts
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import type { MenuGridCellContextValues, MenuGridCellState } from './MenuGridCell.types';
-
-export function useMenuGridCellContextValues_unstable(state: MenuGridCellState): MenuGridCellContextValues {
-  const menuGridCell = React.useMemo(() => ({}), []);
-
-  return { menuGridCell };
-}