+++ packages/react-components/react-menu-grid-preview/library/src/components/MenuGridRow/index.ts
@@ -1,11 +1,5 @@
-export type {
-  MenuGridRowContextValues,
-  MenuGridRowProps,
-  MenuGridRowSlots,
-  MenuGridRowState,
-} from './MenuGridRow.types';
+export type { MenuGridRowProps, MenuGridRowSlots, MenuGridRowState } from './MenuGridRow.types';
 export { MenuGridRow } from './MenuGridRow';
 export { renderMenuGridRow_unstable } from './renderMenuGridRow';
 export { useMenuGridRow_unstable } from './useMenuGridRow';
-export { useMenuGridRowContextValues_unstable } from './useMenuGridRowContextValues';
 export { menuGridRowClassNames, useMenuGridRowStyles_unstable } from './useMenuGridRowStyles.styles';