+++ packages/react-components/react-menu-grid-preview/library/src/components/MenuGridRow/renderMenuGridRow.tsx
@@ -1,19 +1,14 @@
 /** @jsxRuntime automatic */
 /** @jsxImportSource @fluentui/react-jsx-runtime */
 import { assertSlots } from '@fluentui/react-utilities';
-import { MenuGridRowContextValues, MenuGridRowSlots, MenuGridRowState } from './MenuGridRow.types';
-import { MenuGridRowContextProvider } from '../../contexts/menuGridRowContext';
+import { MenuGridRowSlots, MenuGridRowState } from './MenuGridRow.types';
 
 /**
  * Redefine the render function to add slots. Reuse the menugroup structure but add
  * slots to children.
  */
-export const renderMenuGridRow_unstable = (state: MenuGridRowState, contextValues: MenuGridRowContextValues) => {
+export const renderMenuGridRow_unstable = (state: MenuGridRowState) => {
   assertSlots<MenuGridRowSlots>(state);
 
-  return (
-    <MenuGridRowContextProvider value={contextValues.menuGridRow}>
-      <state.root />
-    </MenuGridRowContextProvider>
-  );
+  return <state.root />;
 };